+++ AspNetFinalProject/Controllers/Card/Api/CardApiController.cs
@@ -1,6 +1,7 @@
 ﻿using System.Security.Claims;
 using AspNetFinalProject.Common;
 using AspNetFinalProject.DTOs;
+using AspNetFinalProject.Enums;
 using AspNetFinalProject.Mappers;
 using AspNetFinalProject.Services.Interfaces;
 using Microsoft.AspNetCore.Authorization;
@@ -15,12 +16,15 @@ public class CardApiController : ControllerBase
 {
     private readonly ICardService _cardService;
     private readonly ICurrentUserService _currentUserService;
+    private readonly IBoardListService _boardListService;
     
     public CardApiController(ICardService cardService,
-                             ICurrentUserService currentUserService)
+                             ICurrentUserService currentUserService,
+                             IBoardListService boardListService)
     {
         _cardService = cardService;
         _currentUserService = currentUserService;
+        _boardListService = boardListService;
     }
     
     [HttpGet("list/{boardListId:guid}")]
@@ -28,6 +32,7 @@ public class CardApiController : ControllerBase
     {
         var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
         if (userId == null) return Unauthorized();
+        
         var cards = await _cardService.GetCardsByListAsync(boardListId, userId);
         var result = cards.Select(CardMapper.CreateDto);
         return Ok(result);
@@ -50,6 +55,12 @@ public class CardApiController : ControllerBase
         var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
         if (userId == null) return Unauthorized();
 
+        var list = await _boardListService.GetByIdAsync(Guid.Parse(dto.BoardListId));
+        if (list == null) return NotFound();
+        var hasPermission = await _currentUserService.HasBoardRoleAsync(list.BoardId, ParticipantRole.Admin,
+            ParticipantRole.Member, ParticipantRole.Owner);
+        if (!hasPermission) return Forbid();
+        
         var card = await _cardService.CreateAsync(userId, dto);
 
         return CreatedAtAction(nameof(GetCardsByList), new { boardListId = dto.BoardListId }, CardMapper.CreateDto(card));
@@ -75,6 +86,14 @@ public class CardApiController : ControllerBase
         var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
         if (userId == null) return Unauthorized();
 
+        var card = await _cardService.GetByIdAsync(id);
+        if(card == null) return NotFound();
+        var list = await _boardListService.GetByIdAsync(card.BoardListId);
+        if (list == null) return NotFound();
+        var hasPermission = await _currentUserService.HasBoardRoleAsync(list.BoardId, ParticipantRole.Admin,
+            ParticipantRole.Member, ParticipantRole.Owner);
+        if (!hasPermission) return Forbid();
+        
         var deleted = await _cardService.DeleteAsync(id, userId);
         if (!deleted) return NotFound();
